# cmd/api/docs/swagger.yaml
definitions:
...
  schemas.PointSaleResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  schemas.Response:
    properties:
      body: {}
      message:
        type: string
      status:
        type: boolean
    type: object
  schemas.RoleResponse:
    properties:
      id:
        type: integer
      name:
        enum:
        - admin
        - vendedor
        - repositor
        type: string
    type: object
  schemas.UserCreate:
    properties:
      address:
        type: string
      cellphone:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      point_sales_ids:
        items:
          type: integer
        type: array
      role_id:
        type: integer
      username:
        type: string
    type: object
  schemas.UserResponse:
    properties:
      address:
        type: string
      cellphone:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      last_name:
        type: string
      point_sales:
        items:
          $ref: '#/definitions/schemas.PointSaleResponse'
        type: array
      role:
        $ref: '#/definitions/schemas.RoleResponse'
      username:
        type: string
    type: object
  schemas.UserResponseDTO:
    properties:
      address:
        type: string
      cellphone:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      is_admin:
        type: boolean
      last_name:
        type: string
      role:
        $ref: '#/definitions/schemas.RoleResponse'
      username:
        type: string
    type: object
  schemas.UserUpdate:
    properties:
      address:
        type: string
      cellphone:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      point_sales_ids:
        items:
          type: integer
        type: array
      role_id:
        type: integer
      username:
        type: string
    type: object
info:
  contact: {}
  description: This is a API for Club Norte
  termsOfService: http://swagger.io/terms/
  title: APP Club Norte
  version: "1.0"
paths:
...
  /v1/user/create:
    post:
      consumes:
      - application/json
      description: UserCreate Crear un nuevo usuario
      parameters:
      - description: user create
        in: body
        name: user_create
        required: true
        schema:
          $ref: '#/definitions/schemas.UserCreate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: UserCreate
      tags:
      - User
  /v1/user/delete/{id}:
    delete:
      consumes:
      - application/json
      description: UserDelete Eliminar un usuario por ID
      parameters:
      - description: Id del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: UserDelete
      tags:
      - User
  /v1/user/get/{id}:
    get:
      consumes:
      - application/json
      description: UserGetByID obtener usuario por ID
      parameters:
      - description: Id del usuario
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: UserGetByID
      tags:
      - User
  /v1/user/get_all:
    get:
      consumes:
      - application/json
      description: UserGetAll obtener todos los usuarios
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/schemas.UserResponseDTO'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: UserGetAll
      tags:
      - User
  /v1/user/get_by_email:
    get:
      consumes:
      - application/json
      description: UserGetByEmail obtener usuario por Email
      parameters:
      - description: email del usuario
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/schemas.Response'
            - properties:
                body:
                  $ref: '#/definitions/schemas.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: UserGetByEmail
      tags:
      - User
  /v1/user/update:
    put:
      consumes:
      - application/json
      description: UserUpdate Editar un usuario
      parameters:
      - description: user create
        in: body
        name: user_update
        required: true
        schema:
          $ref: '#/definitions/schemas.UserUpdate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/schemas.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/schemas.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/schemas.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/schemas.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/schemas.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/schemas.Response'
      security:
      - CookieAuth: []
      summary: UserUpdate
      tags:
      - User
securityDefinitions:
  CookieAuth:
    description: 'Enter your JWT token here. Example: "eyJhbGciOiJIUz..."'
    in: cookie
    name: access_token
    type: apiKey
swagger: "2.0"